+++ src/ast/parser.rs
@@ -400,7 +400,7 @@ impl Parser {
         match &token.kind {
             TokenKind::Number(number) => {
                 self.consume();
-                Ok(Expression::new(ExpressionKind::Number(NumberExpression::new(number.clone() as i64))))
+                Ok(Expression::new(ExpressionKind::Number(NumberExpression::new(number.clone() as i32))))
             },
             TokenKind::Word(word) => {
                 self.consume();
